Move `samplers` module to root module
