+++ docs/api/samplers.md
@@ -1,1 +1,1 @@
-::: outlines.generate.samplers
+::: outlines.samplers
