+++ outlines/generate/choice.py
@@ -1,8 +1,8 @@
 from functools import singledispatch
 from typing import Callable, List, Optional
 
-from outlines.generate.samplers import Sampler, multinomial
 from outlines.models import OpenAI
+from outlines.samplers import Sampler, multinomial
 
 from .regex import regex
 
@@ -12,7 +12,7 @@ def choice(
     model,
     choices: List[str],
     max_tokens: Optional[int] = None,
-    sampler: Sampler = multinomial,
+    sampler: Sampler = multinomial(),
 ):
     regex_str = r"(" + r"|".join(choices) + r")"
     return regex(model, regex_str, max_tokens, sampler)
@@ -23,9 +23,9 @@ def choice_openai(
     model: OpenAI,
     choices: List[str],
     max_tokens: Optional[int] = None,
-    sampler: Sampler = multinomial,
+    sampler: Sampler = multinomial(),
 ) -> Callable:
-    if not sampler == multinomial:
+    if not isinstance(sampler, multinomial):
         raise NotImplementedError(
             r"The OpenAI API does not support any other sampling algorithm "
             + "that the multinomial sampler."
