+++ tests/test_samplers.py
@@ -2,12 +2,7 @@ import math
 
 import torch
 
-from outlines.generate.samplers import (
-    GreedySampler,
-    MultinomialSampler,
-    greedy,
-    multinomial,
-)
+from outlines.samplers import GreedySampler, MultinomialSampler, greedy, multinomial
 
 
 def test_aliases():
@@ -21,18 +16,10 @@ def test_greedy():
     next_token_ids = sampler(logits)
     assert next_token_ids.equal(torch.tensor([[2]]))
 
-    sampler = GreedySampler(2)
-    next_token_ids = sampler(logits)
-    assert next_token_ids.equal(torch.tensor([[2], [1]]))
-
     sampler = GreedySampler()
     logits = torch.tensor([[10.0, 0.0, 3.0], [-math.inf, 2.0, 5.0]])
     next_token_ids = sampler(logits)
     assert next_token_ids.equal(torch.tensor([[0], [2]]))
-
-    sampler = GreedySampler(2)
-    next_token_ids = sampler(logits)
-    assert next_token_ids.equal(torch.tensor([[0, 2], [2, 1]]))
 
 
 def test_multinomial():
